general refactoring
- split source in multiple modules
