+++ src/pop3_stat.rs
@@ -1,0 +1,8 @@
+/// POP3 maildrop statistics
+pub struct Pop3Stat {
+    /// count of massages in the maildrop
+    pub message_count: u32,
+
+    /// size of the maildrop in bytes
+    pub maildrop_size: u32,
+}
